Split out Solidity-specific part of ExecutionFramework
